search thing done

+++ lib/src/modules/home/ui/search.dart
@@ -75,6 +75,11 @@ class _SearchPageState extends State<SearchPage> {
     length < 1 ? isDBListEmpty = true : isDBListEmpty = false;
   }
 
+  getListData() async {
+    List listItem = await HiveDbServices().getListData(Strings.hiveLogName);
+    return listItem;
+  }
+
   deleteItem() async {
     int itemcount = await HiveDbServices().getListLength(Strings.hiveLogName);
     if (itemcount > 5) {
@@ -298,9 +303,12 @@ class _SearchPageState extends State<SearchPage> {
                       padding: EdgeInsets.only(bottom: 20),
                       scrollDirection: Axis.vertical,
                       itemCount:
-                          snapshot.data.length < 5 ? snapshot.data.length : 5,
+                          snapshot.data.length < 10 ? snapshot.data.length : 10,
                       itemBuilder: (BuildContext context, int index) {
-                        return _buildRecentItem(index, snapshot.data);
+                        List reverseList = List.from(snapshot.data.reversed);
+                        print(reverseList);
+                        // return _buildRecentItem(index, snapshot.data);
+                        return _buildRecentItem(index, reverseList);
                       },
                     ),
                   )
@@ -424,38 +432,47 @@ class _SearchPageState extends State<SearchPage> {
                                               .primaryVariant)),
                             ),
                             onTap: () async {
-                              _route(data);
-                              print(Recent);
-                              if (data != null) {
-                                deleteItem();
-                                final recentitem = Recent(
-                                    1,
-                                    data.titleC,
-                                    data.appIconUrlC,
-                                    data.id,
-                                    data.name,
-                                    data.objectName,
-                                    data.rtfHTMLC,
-                                    data.typeC,
-                                    // data.schoolId,
-                                    // data.dept,
-                                    data.statusC,
-                                    data.urlC,
-                                    data.pdfURL,
-                                    data.sortOrder,
-                                    data.deepLink,
-                                    data.appURLC,
-                                    data.calendarId,
-                                    data.emailC,
-                                    data.imageUrlC,
-                                    data.phoneC,
-                                    data.webURLC,
-                                    data.address,
-                                    data.geoLocation,
-                                    data.descriptionC,
-                                    data.latitude,
-                                    data.longitude);
-                                addtoDataBase(recentitem);
+                              List itemListData = await getListData();
+                              List idList = [];
+                              for (int i = 0; i < itemListData.length; i++) {
+                                idList.add(itemListData[i].id);
+                              }
+
+                              if (idList.contains(data.id)) {
+                              } else {
+                                _route(data);
+                                print(Recent);
+                                if (data != null) {
+                                  deleteItem();
+                                  final recentitem = Recent(
+                                      1,
+                                      data.titleC,
+                                      data.appIconUrlC,
+                                      data.id,
+                                      data.name,
+                                      data.objectName,
+                                      data.rtfHTMLC,
+                                      data.typeC,
+                                      // data.schoolId,
+                                      // data.dept,
+                                      data.statusC,
+                                      data.urlC,
+                                      data.pdfURL,
+                                      data.sortOrder,
+                                      data.deepLink,
+                                      data.appURLC,
+                                      data.calendarId,
+                                      data.emailC,
+                                      data.imageUrlC,
+                                      data.phoneC,
+                                      data.webURLC,
+                                      data.address,
+                                      data.geoLocation,
+                                      data.descriptionC,
+                                      data.latitude,
+                                      data.longitude);
+                                  addtoDataBase(recentitem);
+                                }
                               }
                             }),
                       );
@@ -492,8 +509,7 @@ class _SearchPageState extends State<SearchPage> {
       return CustomIconWidget(
         iconUrl: obj.appIconUrlC ?? Overrides.defaultIconUrl,
       );
-    }
-    else if (obj.appIconC != null) {
+    } else if (obj.appIconC != null) {
       return Icon(
         IconData(
           int.parse('0x${obj.appIconC!}'),
@@ -503,8 +519,7 @@ class _SearchPageState extends State<SearchPage> {
         color: Theme.of(context).colorScheme.primary,
         size: Globals.deviceType == "phone" ? 24 : 32,
       );
-    }
-    else {
+    } else {
       return CustomIconWidget(
         iconUrl: Overrides.defaultIconUrl,
       );
